[feat] URL extension 구현
